major: text+url parser; rss/atom types; json/yaml/markdown outputs; add dry-run
this breaks the state file and config: content changed, and now we have xpath_*

+++ urlobs.py
@@ -1,19 +1,26 @@
 #!/usr/bin/env python3
-import requests
-import yaml
-import hashlib
+import sys, json, yaml, hashlib, time, re, requests, traceback
 from lxml import etree
 from bs4 import BeautifulSoup
-import json
-import time
-import re
-import traceback
+from dataclasses import dataclass, asdict
+from urllib.parse import urljoin
 from argparse import ArgumentParser
 
 
+def safe_first(xs):
+    if len(xs) == 0: return None
+    else: return xs[0]
+
 def process_content(text):
+    text = re.sub(r'^ *[-—] *', '', text)
     # Replace sequences of spaces, tabs, and carriage returns with a single space
-    return re.sub(r'[ \t\r]+', ' ', text)
+    text = re.sub(r'[ \t\r]+', ' ', text)
+    return text
+
+def escape_chars(input: str, special: str) -> str:
+    for char in special:
+        input = input.replace(char, f'\\{char}')
+    return input
 
 def fetch_content(url, headers=None, post_data=None, verbose=False):
     # Fetch content from URL using GET or POST
@@ -26,39 +33,110 @@ def fetch_content(url, headers=None, post_data=None, verbose=False):
     response.raise_for_status()
     return response.text
 
-def parse_content(page, content_type, xpath, fn=None):
+@dataclass(frozen=True, order=True)
+class Item:
+    text: str
+    url: str
+
+yaml.add_representer(Item, lambda dumper, data: dumper.represent_dict(asdict(data)),
+                     Dumper=yaml.SafeDumper)
+
+class JsonEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, Item):
+            return asdict(obj)
+        return super().default(obj)
+
+def parse_content(page, content_type, xpath, xpath_text, xpath_url, fn):
     # Parse content based on type (HTML, XML, JSON)
     if content_type == "html":
         soup = BeautifulSoup(page, 'html.parser')
-        return [ element.get_text() for element in soup.select(xpath) ]
-    elif content_type == "xml":
+        def get_text(element):
+            if xpath_text and (el := safe_first(element.select(xpath_text))):
+                return el.get_text()
+            else:
+                return element.get_text()
+        def get_url(element):
+            if xpath_url and (el := safe_first(element.select(xpath_url))):
+                if (href := el.get('href') or el.get('src')):
+                    return href
+                else:
+                    return el.get_text()
+            elif (href := element.get('href') or element.get('src')):
+                return href
+            elif (el := element.find(attrs={'href': True}) or element.find(attrs={'src': True})):
+                return el.get('href') or el.get('src')
+        return [ dict(text=get_text(element), url=get_url(element))
+                for element in soup.select(xpath) ]
+
+    elif content_type in ("xml", "rss", "atom"):
+        if content_type == 'atom' and 'xmlns' in page[:500]:
+            page = re.sub(r'xmlns="[^"]+"', "", page) # F namespaces
         tree = etree.fromstring(page.encode('utf8'))
-        return [ element.text for element in tree.xpath(xpath) ]
+
+        def get_text(element):
+            if xpath_text and (el := safe_first(element.xpath(xpath_text))) is not None:
+                return el.text
+            else:
+                return element.text
+        def get_url(element):
+            if xpath_url and (el := safe_first(element.xpath(xpath_url))) is not None:
+                if (href := el.get('href') or el.get('src')):
+                    return href
+                else:
+                    return el.text
+            elif (href := element.get('href') or element.get('src')):
+                return href
+            elif (el := element.xpath('*[@href and @href!=""]') or element.xpath('*[@src and @src!=""]')):
+                return el.get('href') or el.get('src')
+        return [ dict(text=get_text(element), url=get_url(element))
+                for element in tree.xpath(xpath) ]
+
     elif content_type == "json":
         data = json.loads(page)
-        return eval(fn, {}, {'data': data})
+        xs = eval(fn, {}, {'data': data}) # arbitrary code execution
+        return list(xs)
+
     else:
         raise ValueError(f"Unknown content type: {content_type}")
 
-def main(config='url.yaml', verbose=False):
+def main(config='url.yaml', verbose=False, format='text', dry_run=False):
     with open(config, 'r', encoding='utf-8') as file:
         urls = yaml.safe_load(file)
+
+    outputs = {}
 
     for info in urls:
         url = info.get('url')
         if verbose: print(f"processing {url}")
-        post_data = info.get('post') # POST data, implies POST request
-        xpath = info.get('xpath')
-        fn = info.get('fn')
+
         content_type = info.get('type', 'html')
+        xpath = None
+        xpath_text = None
+        xpath_url = None
+        if content_type == 'rss':
+            xpath = '/rss/channel/item'
+            xpath_text = './title'
+            xpath_url = './link'
+        elif content_type == 'atom':
+            xpath = '/feed/entry'
+            xpath_text = './title'
+            xpath_url = './link'
+
+        post_data = info.get('post') # POST data, implies POST request
+        xpath = info.get('xpath', xpath)
+        xpath_text = info.get('xpath_text', xpath_text)
+        xpath_url = info.get('xpath_url', xpath_url)
+        fn = info.get('fn')
         title = info.get('title', url)
         hash_value = info.get('hash', '')
-        old_content = info.get('content', [])
+        old_items = [ Item(**x) for x in info.get('content', []) ]
         freq = info.get('interval', 0)
         last_date = info.get('last', 0)
         keep_old = info.get('keep_old', 0)
-        no_order = info.get('no_order', keep_old)
+        no_order = info.get('no_order', 0)
         only_diffs = info.get('onlydiffs', 1)
+        only_news = info.get('onlynews', 0)
         failures = info.get('failures', 0)
         user_agent = info.get('user_agent')
         cookie = info.get('cookie')
@@ -82,36 +160,44 @@ def main(config='url.yaml', verbose=False):
             if verbose: print(f"fetching {url}, {headers=}")
             page = fetch_content(url, headers=headers, post_data=post_data, verbose=verbose)
             info['last'] = now
-            render = parse_content(page, content_type, xpath, fn)
-            if keep_old:
-                render.extend(( el for el in old_content if el not in render ))
+            xs = parse_content(page, content_type, xpath, xpath_text, xpath_url, fn)
+            items = [ Item(text=process_content(x['text']),
+                           url=urljoin(url, x['url'].strip()))
+                     for x in xs ]
+
+            if keep_old > 0:
+                items.extend([ el
+                               for el in old_items
+                               if el not in items ][:keep_old])
             if no_order:
-                render.sort()
-            render = [ process_content(el) for el in render ]
+                items.sort()
 
             if verbose:
-                print(f"old: {old_content}")
-                print(f"rendered: {render}")
+                print(f"old: {old_items}")
+                print(f"rendered: {items}")
 
-            new_hash = hashlib.md5(''.join(render).encode('utf-8')).hexdigest()
+            new_hash = hashlib.md5(str(items).encode('utf-8')).hexdigest()
             if verbose: print(f"hashed: {new_hash}")
 
             if hash_value != new_hash:
-                diffs = []
+                output = {}
                 if only_diffs:
-                    old_set, new_set = set(old_content), set(render)
-                    diffs.append("++ New:")
-                    diffs.extend(( f"  {el}" for el in new_set - old_set ))
-                    diffs.append("-- Off:")
-                    diffs.extend(( f"  {el}" for el in old_set - new_set ))
-                else:
-                    diffs = list(set(old_content) ^ set(render))
-                print(f"〉 Changes for {title}:\n" + "\n".join(diffs) + "\n")
+                    old_set = set(old_items)
+                    new_set = set(items)
+                    if (new := new_set - old_set):
+                        output['new'] = [ x for x in items if x in new ] # keep order
+                    if not only_news and (old := old_set - new_set):
+                        output['old'] = [ x for x in items if x in old ] # keep order
+                elif (all := list(dict.fromkeys(items))): # dedup
+                    output['all'] = all
+
+                if output: # only if not empty
+                    outputs[title] = output
             elif verbose:
                 print(f"No change for {title}")
 
             info['hash'] = new_hash
-            info['content'] = render
+            info['content'] = items
             info['failures'] = 0
         except Exception as e:
             if verbose: traceback.print_exception(e)
@@ -120,13 +206,53 @@ def main(config='url.yaml', verbose=False):
             if failures >= min_alert_failures:
                 print(f"〉✗ Alert: Fetch failed for {title} (freq: {freq})")
 
-    with open(config, 'w', encoding='utf-8') as file:
-        yaml.safe_dump(urls, file, sort_keys=False)
+    # outputs
+    if format == 'text':
+        for title, output in outputs.items():
+                print(f"〉 Changes for {title}:\n")
+                def print_them(xs):
+                    print('\n'.join(( f"  {x.text}" for x in xs )))
+                if (new := output.get('new')):
+                    print("++ New:")
+                    print_them(new)
+                if (old := output.get('old')):
+                    print("-- Old:")
+                    print_them(old)
+                if (all := output.get('all')):
+                    print_them(all)
+
+    if format == 'markdown':
+        for title, output in outputs.items():
+                print(f"# {title}\n")
+                def print_them(xs):
+                    print('\n'.join(( f"- [{escape_chars(x.text, '[]')}]({escape_chars(x.url, '()')})"
+                                     for x in xs )))
+                if (new := output.get('new')):
+                    print("## New")
+                    print_them(new)
+                if (old := output.get('old')):
+                    print("## Old")
+                    print_them(old)
+                if (all := output.get('all')):
+                    print_them(all)
+
+    elif format == 'json':
+        json.dump(outputs, sys.stdout, cls=JsonEncoder)
+
+    elif format == 'yaml':
+        yaml.safe_dump(outputs, sys.stdout)
+
+    # write state back to file
+    if not dry_run:
+        with open(config, 'w', encoding='utf-8') as file:
+            yaml.safe_dump(urls, file, sort_keys=False)
 
 if __name__ == "__main__":
     parser = ArgumentParser()
     parser.add_argument('config')
     parser.add_argument('--verbose', default=False, action='store_true')
+    parser.add_argument('--format', default='text', choices=('text', 'markdown', 'json', 'yaml'))
+    parser.add_argument('--dry-run', default=False, action='store_true')
     args = parser.parse_args()
     if args.verbose: print(f"args: {args}")
     main(**vars(args))
